Refactor index.js to call app.js inside src
